# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Subscription API With Swagger",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/subs/total": {
            "post": {
                "description": "Calculate total price for subscriptions in given period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Calculate total price",
                "parameters": [
                    {
                        "description": "Period and filter criteria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SumRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Total price calculation result",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid date format or period",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/subs/{id}": {
            "delete": {
                "description": "Mark subscription as deleted (soft delete)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Delete subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Subscription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sub {id} was deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "Subscription not found",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.SumRequest": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}